controllers set up complete
