+++ backend/src/Controller/Api/Auth/AuthController.php
@@ -1,18 +1,18 @@
 <?php
 
-namespace App\Controller;
+namespace App\Controller\Api\Auth;
 
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 
-final class OrderController extends AbstractController
+final class AuthController extends AbstractController
 {
-    #[Route('/order', name: 'app_order')]
+    #[Route('/auth', name: 'app_auth')]
     public function index(): Response
     {
-        return $this->render('order/index.html.twig', [
-            'controller_name' => 'OrderController',
+        return $this->render('auth/index.html.twig', [
+            'controller_name' => 'AuthController',
         ]);
     }
 }
