+++ backend/src/Controller/Api/Menu/MenuController.php
@@ -1,18 +1,18 @@
 <?php
 
-namespace App\Controller;
+namespace App\Controller\Api\Menu;
 
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 
-final class UserController extends AbstractController
+final class MenuController extends AbstractController
 {
-    #[Route('/user', name: 'app_user')]
+    #[Route('/menu', name: 'app_menu')]
     public function index(): Response
     {
-        return $this->render('user/index.html.twig', [
-            'controller_name' => 'UserController',
+        return $this->render('menu/index.html.twig', [
+            'controller_name' => 'MenuController',
         ]);
     }
 }
