+++ backend/src/Controller/Api/Order/OrderController.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace App\Controller\Api\Order;
+
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Attribute\Route;
+
+final class OrderController extends AbstractController
+{
+    #[Route('/order', name: 'app_order')]
+    public function index(): Response
+    {
+        return $this->render('order/index.html.twig', [
+            'controller_name' => 'OrderController',
+        ]);
+    }
+}
