--- backend/src/Controller/AuthController.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace App\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Attribute\Route;
-
-final class AuthController extends AbstractController
-{
-    #[Route('/auth', name: 'app_auth')]
-    public function index(): Response
-    {
-        return $this->render('auth/index.html.twig', [
-            'controller_name' => 'AuthController',
-        ]);
-    }
-}
