--- backend/templates/auth/index.html.twig
@@ -1,20 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Hello AuthController!{% endblock %}
-
-{% block body %}
-<style>
-    .example-wrapper { margin: 1em auto; max-width: 800px; width: 95%; font: 18px/1.5 sans-serif; }
-    .example-wrapper code { background: #F5F5F5; padding: 2px 6px; }
-</style>
-
-<div class="example-wrapper">
-    <h1>Hello {{ controller_name }}! ✅</h1>
-
-    This friendly message is coming from:
-    <ul>
-        <li>Your controller at <code>/Users/dev/Desktop/abraham/snap-serve/backend/src/Controller/AuthController.php</code></li>
-        <li>Your template at <code>/Users/dev/Desktop/abraham/snap-serve/backend/templates/auth/index.html.twig</code></li>
-    </ul>
-</div>
-{% endblock %}
